Clean and organize submit-default-delivery-form e2e test - The file was deleted and the scenarios were included in pregnancy-complete-a-delivery test
