+++ tests/page-objects/default/enketo/default-delivery-report.wdio.page.js
@@ -1,32 +1,9 @@
-const utils = require('@utils');
-const fs = require('fs');
 const moment = require('moment');
 const genericForm = require('./generic-form.wdio.page');
 const reportsPage = require('../reports/reports.wdio.page');
 
-const xml = fs.readFileSync(`${__dirname}/../../../../config/default/forms/app/delivery.xml`, 'utf8');
 const formId = 'delivery';
-
-const docs = [
-  {
-    _id: 'form:dd',
-    internalId: 'DD',
-    title: 'Default Delivery',
-    type: 'form',
-    _attachments: {
-      xml: {
-        content_type: 'application/octet-stream',
-        data: Buffer.from(xml).toString('base64'),
-      },
-    },
-  },
-];
-
 const getYNValue = (boolean) => boolean ? 'yes' : 'no';
-
-const selectPatientName = (name) => {
-  return genericForm.selectContact(name);
-};
 
 const selectNoOfBabiesDelivered = async (value) => {
   const field = await $(`[name="/${formId}/delivery_outcome/babies_delivered_other"]`);
@@ -47,6 +24,7 @@ const populateDeadBabyInformation = async (index, data = { place: 'health_facili
   const dateOfDeathPicker = await section.$(`[name="${repeatPath}/baby_death_date"]`);
   const date = moment(data.date).format('YYYY-MM-DD');
   await reportsPage.setDateInput(dateOfDeathPicker, date);
+  await genericForm.formTitle().click();
   await (await section.$(`[data-name="${repeatPath}/baby_death_place"][value="${data.place}"]`)).click();
   await (await section.$(`[data-name="${repeatPath}/stillbirth"][value="${getYNValue(data.stillbirth)}"]`)).click();
   await (await section.$(`[name="${repeatPath}/baby_death_add_notes"]`)).setValue(`Baby ${index} death`);
@@ -63,6 +41,7 @@ const populateAliveBabyInformation = async (index, data = { sex: 'male', danger:
 
   await (await section.$(`[data-name="${repeatPath}/baby_condition"][value="alive_well"]`)).click();
   await (await section.$(`[name="${repeatPath}/baby_name"]`)).setValue(`AliveBaby-${index}`);
+  await genericForm.formTitle().click();
   await (await section.$(`[data-name="${repeatPath}/baby_sex"][value="${data.sex}"]`)).click();
   await (await section.$(`[data-name="${repeatPath}/birth_weight_know"][value="no"]`)).click();
   await (await section.$(`[data-name="${repeatPath}/birth_length_know"][value="no"]`)).click();
@@ -86,29 +65,8 @@ const populateAliveBabyInformation = async (index, data = { sex: 'male', danger:
   }
 };
 
-const getDeadBabyUUID = async (index) => {
-  const element = await $(
-    `//*[text()="report.DD.baby_death.baby_death_repeat.${index}.baby_death_profile_doc"]/../../p`
-  );
-  return await element.getText();
-};
-
-const getAliveBabyUUID = async (index) => {
-  const element = await $(
-    `//*[text()="report.DD.babys_condition.baby_repeat.${index}.baby_details.child_doc"]/../../p`
-  );
-  return await element.getText();
-};
-
 module.exports = {
-  formInternalId: docs[0].internalId,
-  configureForm: (userContactDoc) => {
-    return utils.seedTestData(userContactDoc, docs);
-  },
-  selectPatientName,
   selectNoOfBabiesDelivered,
   populateDeadBabyInformation,
   populateAliveBabyInformation,
-  getDeadBabyUUID,
-  getAliveBabyUUID,
 };
